+++ source/aogmaneo/image_encoder.h
@@ -112,9 +112,9 @@ public:
     }
 
     // serialization
-    int size() const; // returns size in bytes
-    int state_size() const; // returns state size in bytes
-    int weights_size() const; // returns weights size in bytes
+    long size() const; // returns size in bytes
+    long state_size() const; // returns state size in bytes
+    long weights_size() const; // returns weights size in bytes
 
     void write(
         Stream_Writer &writer
